openclaw-skill/requirements.txt
# AgentMesh OpenClaw Skill Dependencies

# Cryptography
pynacl>=1.5.0
cryptography>=41.0.0

# Networking
websockets>=12.0
aiohttp>=3.9.0

# Optional: Matrix Olm for full Signal Protocol
# python-olm>=3.2.0

# Utilities
